feat: command with bang runs in terminal mode (no ui)
This is to allow for running any command in terminal mode. A use case is
when running git diff; it looks better with terminal mode when using
something like `delta`, but it's more vim-like and integrated when using
an actual buffer. So this way a user can easily decide what they want
per command invocation.