fix: issue where popup keymap wouldn't close properly
The issue was that we were attempting to navigate to the previous buffer
before closing the popup buffer, as that is the default behavior of
`deregister_buffer`. Renamed the flag to skip this to
`skip_go_to_last_buffer` and used this flag to not use this behavior for
the popup buffer.