feat: make some buffers listed when created
This is to make it clear that these buffers need to be closed, so they
don't hang around causing performance issues. Also, this means that
pressing "q" in some contexts will go back to the last ever buffer
instead of the last non-ever buffer. I think this is what we want, but
not sure.