fix: bug where `:G branch -D` caused UI issues
The underlying problem was that any branch command used the branch UI,
including `:G branch -D`. This created a buffer with the text "Deleted
branch {some branch name}". Pressing <enter> on this line used the
"log" mapping, which is not what we want.

Changed it up so that branch commands that do anything aside from just
listing the branches will skip the UI and go straight to terminal
mode instead.