feat: update deregister behavior to be smarter
If there is a non-Trunks buffer associated with the window when closing
a Trunks buffer, go back to that buffer before closing all Trunks
buffers for that window. Otherwise, just close all the Trunks buffers.
This means splits are now always closed if they were made by Trunks.

Worth keeping an eye on this in case there is some unexpected behavior.