bug: fix broken difftool
The issue was that we switched up `stream_lines`, so that sometimes
lines wouldn't render until `on_exit` was called. However, if `silent`
is passed to `stream_lines`, `on_exit` is never called, which was
causing the staging area to never render lines. Moving the `silent`
logic to a more appropriate location fixes the issue.